requirements.txt
pandas
openpyxl
sentence-transformers
transformers
faiss-cpu
torch
yfinance
streamlit